Add elevator planner and tests
This revision of the elevator planner is very simple. It
tries to obey a few rules:

1. Dont let the elevator or wrist go past their limits
2. Don't allow the elevator to move a long distance while the wrist is out.
   Instead, stow the wrist then drive the elevator.
3. Do not violate a minimum angle constraint on the wrist
   when dunking over the top of the elevator
4. Do not even come close to crossing the intake through
   the crossbar.
5. Do not open the wrist when the pivot angle is < 60 degrees. This
   probably requires some smarts in the intake subsystem to enforce,
   but the elevator at least won't move yet.

You can build a planner with the class ElevatorMotionPlanner and
update its desired state (ElevatorState, contains height, pivot angle,
and the probable jaw state), then call update iteratively (designed for
~200Hz) and safely write that output into a Talon position controller.

This is still rought and probably won't look very optimal. But the
elevator shouldn't break, so I guess we've got that going for us.