First pass at auto mode selection
This is sort of complex. The UI to the user is currently pick a
mode. The available modes
are SimpleSwitch, ScaleAndSwitch (quals), ScaleOnly.

For each of these modes we may actually need to select a different
mode (extends AutoModeBase) based on the field state we get
from the DS.

Addtionally, we want to precompute the paths for each.